+++ test/commands/matilda_core/app_memberships/remove_member_command_test.rb
@@ -1,0 +1,43 @@
+# frozen_string_literal: true
+
+require 'test_helper'
+
+module MatildaCore
+
+  module AppMemberships
+
+    # RemoveMemberCommandTest.
+    class RemoveMemberCommandTest < ActiveSupport::TestCase
+
+      test 'il gruppo è obbligatorio' do
+        command = MatildaCore::AppMemberships::RemoveMemberCommand.new(
+          user_uuid: 'general',
+          log_who: ''
+        )
+        assert_not command.completed?
+        assert_command_error_code(command, :group_uuid)
+      end
+
+      test "l'utente è obbligatorio" do
+        command = MatildaCore::AppMemberships::RemoveMemberCommand.new(
+          group_uuid: 'general',
+          log_who: ''
+        )
+        assert_not command.completed?
+        assert_command_error_code(command, :user_uuid)
+      end
+
+      test 'caso positivo' do
+        command = MatildaCore::AppMemberships::RemoveMemberCommand.new(
+          group_uuid: 'general',
+          user_uuid: 'general',
+          log_who: ''
+        )
+        assert command.completed?
+      end
+
+    end
+
+  end
+
+end